binary-search-tree: Moved the solution to the example.rs file
